requirements.txt
django-allauth
openpyxl
